+++ layers/SiteBuilder/packages/component-model-configuration/src/ComponentProcessors/AbstractComponentProcessor.php
@@ -4,17 +4,21 @@ declare(strict_types=1);
 
 namespace PoP\ConfigurationComponentModel\ComponentProcessors;
 
-use PoP\ComponentModel\Component\Component;
 use PoP\ComponentModel\Checkpoints\CheckpointInterface;
+use PoP\ComponentModel\Component\Component;
 use PoP\ComponentModel\ComponentProcessors\AbstractComponentProcessor as UpstreamAbstractComponentProcessor;
 use PoP\ComponentModel\ComponentProcessors\FormattableModuleInterface;
 use PoP\ComponentModel\Constants\DataLoading;
+use PoP\ComponentModel\Constants\FieldOutputKeys;
 use PoP\ComponentModel\Misc\GeneralUtils;
 use PoP\ComponentModel\Settings\SettingsManagerFactory;
 use PoP\ConfigurationComponentModel\Constants\Params;
 use PoP\Definitions\Constants\Params as DefinitionsParams;
+use PoP\GraphQLParser\Spec\Parser\Ast\LeafField;
+use PoP\GraphQLParser\StaticHelpers\LocationHelper;
 use PoP\Root\App;
 use PoP\Root\Feedback\FeedbackItemResolution;
+use SplObjectStorage;
 
 abstract class AbstractComponentProcessor extends UpstreamAbstractComponentProcessor implements ComponentProcessorInterface
 {
@@ -37,6 +41,61 @@ abstract class AbstractComponentProcessor extends UpstreamAbstractComponentProce
         // @todo Fix: this method now returns a SplObjectStorage, yet not adapted here
         if ($outputKeys = $this->getFieldToTypeOutputKeys($component, $props)) {
             $ret['outputKeys'] = $outputKeys;
+        }
+
+        return $ret;
+    }
+
+    /**
+     * Watch out: this function messes up PHPStan! It was moved here from upstream
+     * and not finished, as it's not required for GraphQL.
+     *
+     * @return SplObjectStorage<FieldInterface,string> Key: field output key, Value: self object or relational type output key
+     *
+     * @todo Finish/adapt this function, fix the types for PHPStan
+     */
+    public function getFieldToTypeOutputKeys(Component $component, array &$props): SplObjectStorage
+    {
+        /** @var SplObjectStorage<FieldInterface,string> */
+        $ret = new SplObjectStorage();
+        
+        if ($relationalTypeResolver = $this->getRelationalTypeResolver($component)) {
+            if ($typeOutputKey = $relationalTypeResolver->getTypeOutputKey()) {
+                /**
+                 * Place it under "id" because it is for fetching the current object
+                 * from the DB, which is found through resolvedObject.id
+                 */
+                $idField = new LeafField(
+                    FieldOutputKeys::ID,
+                    null,
+                    [],
+                    [],
+                    LocationHelper::getNonSpecificLocation(),
+                );
+                $ret[$idField] = $typeOutputKey;
+            }
+        }
+
+        // This prop is set for both dataloading and non-dataloading components
+        if ($relationalTypeResolver = $this->getProp($component, $props, 'succeeding-typeResolver')) {
+            foreach ($this->getRelationalComponentFieldNodes($component) as $relationalComponentFieldNode) {
+                // If passing a subcomponent fieldname that doesn't exist to the API, then $subcomponent_typeResolver_class will be empty
+                $typeResolver = $this->getDataloadHelperService()->getTypeResolverFromSubcomponentField($relationalTypeResolver, $relationalComponentFieldNode->getField());
+                if ($typeResolver === null) {
+                    continue;
+                }
+                $ret[$relationalComponentFieldNode->getField()] = $this->getFieldQueryInterpreter()->getTargetObjectTypeUniqueFieldOutputKeys($relationalTypeResolver, $relationalComponentFieldNode->getField());
+            }
+            foreach ($this->getConditionalRelationalComponentFieldNodes($component) as $conditionalRelationalComponentFieldNode) {
+                foreach ($conditionalRelationalComponentFieldNode->getRelationalComponentFieldNodes() as $relationalComponentFieldNode) {
+                    // If passing a subcomponent fieldname that doesn't exist to the API, then $subcomponentTypeResolverClass will be empty
+                    $typeResolver = $this->getDataloadHelperService()->getTypeResolverFromSubcomponentField($relationalTypeResolver, $relationalComponentFieldNode->getField());
+                    if ($typeResolver === null) {
+                        continue;
+                    }
+                    $ret[$relationalComponentFieldNode->getField()] = $this->getFieldQueryInterpreter()->getTargetObjectTypeUniqueFieldOutputKeys($relationalTypeResolver, $relationalComponentFieldNode->getField());
+                }
+            }
         }
 
         return $ret;
