+++ layers/SiteBuilder/packages/component-model-configuration/src/ComponentProcessors/ComponentProcessorInterface.php
@@ -6,6 +6,8 @@ namespace PoP\ConfigurationComponentModel\ComponentProcessors;
 
 use PoP\ComponentModel\Component\Component;
 use PoP\ComponentModel\ComponentProcessors\ComponentProcessorInterface as UpstreamComponentProcessorInterface;
+use PoP\GraphQLParser\Spec\Parser\Ast\FieldInterface;
+use SplObjectStorage;
 
 interface ComponentProcessorInterface extends UpstreamComponentProcessorInterface
 {
@@ -21,4 +23,8 @@ interface ComponentProcessorInterface extends UpstreamComponentProcessorInterfac
     public function getRelevantRoute(Component $component, array &$props): ?string;
     public function getRelevantRouteCheckpointTarget(Component $component, array &$props): string;
     public function getDataloadSource(Component $component, array &$props): ?string;
+    /**
+     * @return SplObjectStorage<FieldInterface,string> Key: field output key, Value: self object or relational type output key
+     */
+    public function getFieldToTypeOutputKeys(Component $component, array &$props): SplObjectStorage;
 }
